pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "quantum-portfolio-optimizer"
version = "1.0.0"
description = "Hybrid quantum-classical portfolio optimization with explainable AI for Spanish retail investors"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Your Name", email = "your.email@university.es"}
]
keywords = [
    "quantum-computing",
    "portfolio-optimization",
    "qaoa",
    "fintech",
    "explainable-ai",
    "langchain"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Office/Business :: Financial :: Investment",
]

dependencies = [
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "yfinance>=0.2.28",
    "duckdb>=0.9.0",
    "qiskit>=1.0.0",
    "qiskit-algorithms>=0.3.0",
    "qiskit-optimization>=0.6.0",
    "cvxpy>=1.4.0",
    "scipy>=1.11.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "plotly>=5.15.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "ruff>=0.0.290",
]
notebooks = [
    "jupyter>=1.0.0",
    "ipykernel>=6.25.0",
    "nbformat>=5.9.0",
]

[project.scripts]
quantum-portfolio = "main:main"

[project.urls]
Homepage = "https://github.com/username/quantum-portfolio-optimizer"
Documentation = "https://github.com/username/quantum-portfolio-optimizer#readme"
Repository = "https://github.com/username/quantum-portfolio-optimizer.git"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "config*"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["src", "config"]

[tool.ruff]
line-length = 100
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=term-missing"
